// GMAOWB/src/app/Models/DemTravail.ts
export class DemTravail{
    NumDem: number;
    UserId: number;
    UserName: string;
    Contact: string;
    Service: string;
    Departement: string;
    DateDem:Date;
    HeureDem:Date;
    Nature:string;
    CodeIntervention: string;
    Consommable:string;
    Description: string;
    CodeEquipement: string;
    DateSouh: Date;
    CentreCout:string;
    Designation: string;
    Status: string;
    TEffect: string;
    NatPanne: string;
    TypePanne: string;
    DateDiag: Date;
    DateDInter: Date;
    DateDiagClo: Date;
    DatePreClo:Date;
    DateClo: Date;
    DateFinT:Date;
    HeureFinT:Date;
    DateValid:Date;
    HeureValid:Date;
    DateValidSys:Date;
    HeureValidSys:Date;
    Reception:string;
    ArretMachine:string;
    MotifValid:string;
    NumEquipement:number;
    CodeEmplacement:string;

}